git: Move setup of Git execution environment into factory
Setup of the Git execution environment is currently hosted in the Gitaly
configuration. This shouldn't be any concern of the configuration-code
though, but only of the Git command factory which requires it to spawn
Git commands.

Move construction of the environment into the command factory. This will
allow more control over it and paves the way for feature-flagged Git
rollouts.