git: Fix missing setup of gitaly-hooks binary
One of our tests is verifying that the `WithRefTxHook()` option sets up
the correct environment variables in the spawned Git command. This test
is currently dependent on whether we use the runtime Git hooks directory
or the Ruby one though because the spawned git-update-ref(1) command
will fail with the Ruby directory hooks, but succeed with the runtime
Git hooks directory.

The root cause is that we don't build the `gitaly-hooks` binary, and as
a result the symlinks pointing to it in the runtime hooks directory will
point to a nonexistent path. Git then doesn't even try to execute hooks,
and as a consequence the command will unexpectedly succeed.

Fix this bug by just skipping hook execution altogether. All we want is
to verify that the environment variables are set up as expected.