Refactor TestSuccessfulResolveConflictsRequestHelper to use the common setup
The other tests in the conflicts package use the same setup function to create the
test service. TestSuccessfulResolveConflictsRequestHelper however uses different
setup as it needs to know commit IDs before it can setup the mock hook manager. This
commit refactors the test to use the common setup functions. This allows for later
creating the test repository via the API.