Create repositories via API in conflicts service tests
This commit changes the conflicts service to create its test repositories
via the API rather than placing them directly to the storage. This allows
Praefect to create the repository metadata records when proxying the calls
if the test are ran with Praefect enabled. With this fixed, the metadata
force creation hack is also disabled for the tests.