Makefile: Allow standalone installation of bundled Git
It is not possible to install Gitaly's bundled Git executables without
also installing Gitaly's own binaries. This is a problem for Omnibus:
for the sake of improved caching it will build Git executables before
building Gitaly such that they don't have to be rebuilt whenever there
is a new Gitaly commit. This is done by just calling "make git", which
installs the complete Git distribution without Gitaly-specific parts.
With the upcoming migration from a Git distribution to bundled Git we'll
also want to install the bundled Git version in this step, which is not
easily done right now.

Add two new build targets `build-bundled-git` and `install-bundled-git`
to allow this usecase.