Allow passing in a ClientConn to CreateRepository
Some tests perform custom connection setup, for example when TLS is
used. To avoid complicating the general case of CreateRepository, this
commit adds an option to provide a pre-established ClientConn. This way
the test can perform whatever it needs to do to setup a connection and
CreateRepository doesn't need to get more complicated.