Disable metadata creation hack in ref service tests
This commit amends the ref service's tests to not rely on Praefect's
metadata force creation hack. The hack creates metadata for repositories
when Praefect first sees a request for the repository. This has potential
to hide problems in tests and doesn't match the production scenario. To
prepare for removing the hack entirely, this commit refactors the tests
to create the test repositories through the API rather than placing them
directly in the storage.

RepositoryService is now set up and registered in the test setup as it's
the service that is responsible for repository creations.

The tests themselves are mostly just updated to use the helpers that use
the API. In TestServer_ListRefs, the branch name was changed from master
to main as Gitaly creates repositories with main as the default branch.