operations: Allow rebase to work if patch has already been applied
Recently it appears there has been an uptick of rebase failures due to
"this patch has already been applied" errors. If a target branch
contains changes that are already in the source branch, and a user
rebases, this failure would cause the rebase to stop.

We now skip these empty commits and allow the rebase to continue. This
is already with squash.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/351017

Changelog: fixed