housekeeping: Strip empty config sections
Historically, we've written a lot of state into the gitconfig to set up
remotes to fetch from. We have stopped doing that since then so that the
only thing we have in there is the "gitlab.fullpath" key which helps to
find out what project a given repository belongs to.

The cleanup logic we had in the past didn't clean up the config
completely though: while keys have been unset, section headers were left
behind even if they were empty and thus had no information anymore. The
result is that we have repositories with tens of thousands of empty
sections, up to a point where some Git commands are in fact dominated by
parsing this configuration.

Fix this by stripping empty config sections when running housekeeping
tasks. Unfortunately, there is no good way to do this via git-config(1),
so we have to resort to manual parsing.

Changelog: performance