housekeeping: Rewrite config cleanup test to be more thorough
Our testcase verifying that we unset certain config entries contains
quite a bunch of logic, but test logic itself is broken. As a result, we
did not catch that we do not unset "remote.*.prune" keys correctly.

Refactor the test to be more direct. Instead of playing tricks with
executing git-config(1) we write our own gitconfig and then verify that
it's got the expected state after having performed housekeeping. This
also shows that we incorrectly retain above config key.