operations: Add tests for rebases with empty commits
When rebasing commits it can happen that a subset of commits one is
about to rebase will become empty because they already exist on the
branch one is rebasing onto. When this happens UserRebaseConfirmable is
currently raising an error, which is unintuitive and ultimately doesn't
make a lot of sense.

Add a testcase which demonstrates this failure.