operations: Deprecate the rebase ID
Back during the days when rebases were implemented via worktrees we
tried to amortize costs of creating the worktrees by reusing them. To
identify a worktree for reuse, callers passed a rebase ID identifying
it, where no two concurrent rebase requests must have used the same ID.

Nowadays though rebases are done in-memory without the use of worktrees,
which is a lot more efficient, especially in big repositories. As a
consequence, the rebase ID doesn't carry any value, and the RPC call
`IsRebaseInProgress()` doesn't exist anymore either. It's thus useless,
but we still require it to be passed.

Stop validating the rebase ID and deprecate the field to phase out
support for it.

Changelog: deprecated