operations: Skip rebasing commits which become empty
When rebasing commits onto a target branch, then it can happen that a
subset of the commits become empty in case they have already been
applied on that branch. UserRebaseConfirmable doesn't handle this right
now and instead returns an error to the caller. This behaviour doesn't
make a lot of sense and is quite unintuitive.

Change behaviour to not return an error anymore in case this happens.

Changelog: fixed