Expand flat paths only for the returned page in GetTreeEntries
GetTreeEntries returns the contents of a tree. For directories, it
flattens the directory hierarchy until it hits a subdirectory which
contains either two directories or a file. This is implemented by
recursively reading the tree via cat-file and checking whether the
flattening should continue. This can be expensive if a tree contains
a lot of subtrees. In production, the RPC has been observed to timeout
with a repository that contains +70000 directories in the root of the
repository. The RPC is doing a lot of unnecessary work as it flattens
the path prior to applying any sort of paging. This commit optimizes
the process by computing flat paths only for the directories in the
returned page. This shouldn't affect the end result, as we only retain
the ordering from git and only rearrange the trees to be at the top of
the results.

With this commit in place, loading the first 1000 entries with page
size of 100 takes about 20 seconds. Previously the RPC would fail to
fetch even a single page within the 30s timeout.

Loading a big repository still doesn't work properly though, as the
front-end calls the logs_tree endpoint in Rails, which ends up calling
GetTreeEntries without paging and thus timing out.

Changelog: performance