operations: Enable use of structured errors in UserMergeBranch
In e542a7d87 (operations: Implement rich errors for UserMergeBranch
access checks, 2021-08-30), we have started to make use of the rich
error model as provided by gRPC in the UserMergeBranch RPC such that we
can provide more information to the caller in error cases. This change
had to be hard-disabled in abd3c46a9 (operations: Force-disable changed
error handling in UserMergeBranch, 2021-08-31) given that it caused
incompatibilities downstream in Rails. Since then, downstream has been
adjusted to be compatible with those changes via igtlab!78366.

Stop force-disabling the change such that the feature flag can be used
to roll out this change to production.