operations: Always respect committer timezone in UserApplyPatch
The UserApplyPatch RPC has traditionally not respected the user's
timezone in the result commit's committer metadata. This has changed
recently such that the correct timezone is set for the committer. This
change was rolled out to production and tested successfully.

Remove the feature flag guarding this new behaviour.

Changelog: fixed