repository: Use skipping negotiation algorithm for replication
When doing reference negotiation, client and server try to come up with
common references. When branches are not at the same commit, the
negotiatior tries to go through the commit chain to come up with older
commits which are in common. This can take quite some time though in
case there have been many new commits on the server side.

Git has thus introduced a new "skipping" negotiation algorithm. This
algorithm has the goal of converging faster by skipping commits in
Fibonacci sequences such that we skip commits increasingly faster. The
downside is that the negotiated packfile is like larger, but the upside
is that we can conclude the reference negotiation phase faster. Given
that we have seen in production that negotiation is what dominates
replication times in repositories with huge references this tradeoff
seems worth it.

Convert our replication to use "skipping" negotiation.

Changelog: performance