gittest: Rename helper which detects the GIT_PROTOCOL
The gittest package exposes a helper `EnableGitProtocolV2Support()`.
Despite its naming, this function doesn't enable any Git protocol at
all. Instead, it creates an intercepting Git command factory which
allows us to detect the Git protocol that's when running the remote side
of a fetch or push.

Rename the function to `NewProtocolDetectingCommandFactory()` to more
clearly indicate what it's doing.