Move tree entry helpers out from commit service
Commit service's package contains helpers for working with tree
entries. These helpers are used also in the repository service.
This creates a cyclic dependency when repository service is setup
in commit service's tests to create repositories via the API. This
commit moves the tree entry helpers out of the commit package so the
cyclic dependency is broken.