Use bare repositories in commit service tests
createRepoWithDivergentBranches is using a worktree to build its
test state. This is unnecessarily complex and makes it slightly more
difficult to create the repository through the API in the helper.
This commit refactors the helper to setup the test state in a bare
repository. createCommits is extracted from the function as an
independent helper so it can be reused across tests. createCommits is
then used also in TestSuccessfulCountComitsRequest to replace its custom
test setup.