Disable metadata creation hack in commit service tests
This commit amends the diff service's tests to not rely on Praefect's
metadata force creation hack. The hack creates metadata for repositories
when Praefect first sees a request for the repository. This has potential
to hide problems in tests and doesn't match the production scenario. To
prepare for removing the hack entirely, this commit refactors the tests
to create the test repositories through the API rather than placing them
directly in the storage.

In some tests, the tested branch was changed from 'master' to 'main' as
Gitaly creates repositories with 'main' as the default branch.