Disable metadata creation hack in smarthttp service tests
This commit amends the smarthttp service's tests to not rely on Praefect's
metadata force creation hack. The hack creates metadata for repositories
when Praefect first sees a request for the repository. This has potential
to hide problems in tests and doesn't match the production scenario. To
prepare for removing the hack entirely, this commit refactors the tests
to create the test repositories through the API rather than placing them
directly in the storage.

Some tests had to be adapted as Praefect is returning different error codes
that Gitaly in some cases.