repository: Use heuristic for incremental repacks in OptimizeRepository
The OptimizeRepository call only ever does full repacks of a repository,
if required. This strategy is lacking though: a repository may continue
to write objects into the repository without ever creating packfiles,
and in that case we'd never perform a complete repack but still end up
with a repository that's inefficiently packed.

Introduce a new heuristic which does an incremental repack. In contrast
to the full repack, the incremental repack will only collect all loose
objects and compress them into a single packfile, and thus it doesn't
scale with the size of the repository but instead with the number of
loose objects that currently exist. The heuristic is thus quite simple:
we do an incremental repack whenever we determine the repository to have
more than a specific number of loose objects, where we have set this
number to 1024 objects.

While we could count objects by just walking through the complete object
directory, this would mean we had to open all of the 256 sharding
directories. While this shouldn't be all that bad typically, it does add
some load to the disk. Instead, we're making use of the fact that Git's
object IDs use a cryptograph hash, and as a result objects should have
uniform distribution across all of the sharding directories. We can thus
easily extrapolate from a single shard by multiplying the number of
objects found in there by the number of shards. This is the same
heuristic as git-gc(1) uses when passing the `--auto` switch, except
that we have a much lower boundary of how many loose objects may exist.

Changelog: changed