repository: Use heuristic to prune objects in OptimizeRepository
While we now try to do incremental repacks when we determine that a
repository has too many loose objects, it could be that those loose
objects are in fact unreachable objects which won't ever get written
into a packfile. As a consequence, we'd try to repack repositories every
time because the number of loose objects never goes down. This isn't
much of a problem: git-repack(1) would exit rather fast in this case.
But it's a shortcoming of the way OptimizeRepository is implemented.

Introduce a new heuristic which prunes unreachable objects when there
are still too many objects after having packed the repository, using
the exact same logic we also use for incremental repacks. If we have
more than this number of objects then we'll try to prune objects.

Note that git-prune(1) has a two-week grace period before deleting
unreachable objects. This guards us against accidentally corrupting the
repository in case concurrently running Git commands still refer to the
objects that are unreachable, even though no reference point to them
anymore. Furthermore, it allows us to restore objects during that grace
period, if required.

There is one exception to this new heuristic: when OptimizeRepository is
executed on an object pool we don't ever want to prune objects, so we
exit early in case we see that it is one. This is a hard requirement
right now: object pools don't have the full view of all referenced
objects for all pool members, so it could be that objects in the pool
are unreachable via its current set of references, but any other member
still requires them. While this shouldn't ever happen because we create
keep-alive references to retain unreachable objects in pools, it is both
safer and also more performant to skip pruning because it should in
theory never prune any objects anyway.

Changelog: changed