Create repositories via API in hook service tests
Hook service is currently creating repositories using test helpers
and not via the API. This commit converts majority of the tests to
create the repos through the API so their state matches better how
the repos will actually be in production.

TestPreReceiveHook_GitlabAPIAccess was not converted to create the
repository through the API as it relies on some setup prior to the
repository being created.