repository: Use heuristic to pack refs in OptimizeRepository
Except for the housekeeping tasks which prune corrupt references,
OptimizeRepository is currently completely ignoring references. As a
result, it could be that the repository is well-packed with regards to
its objects but is a mess with regards to its references.

Git has two ways to store references right now: loose references are
the "default" backend, where every changed reference is written into a
single file. Loose references are inefficient though: reading a single
reference requires us to traverse multiple directories, open the file,
scan its contents and so on. Git thus also knows a packed format of
references, where the git-pack-refs(1) command will collect all loose
refs and write them into a single packed-refs file. Reading this format
is a lot more efficient. We should thus make sure that we regularly
repack refs.

Introduce a new heuristic which packs loose references: whenever there
are too many of them we run git-pack-refs(1) to pack them. Packing refs
scales with the number of references which exist in the repository. We
thus use the size of the existing packed-refs file to inform ourselves
about how long it's likely to take to pack them. The bigger its size,
the more loose refs we accept in the repository. This is a tradeoff that
trades runtime of OptimizeRepository against git-pack-refs(1).

We use a similar heuristic here as for packing objects by using a
logarithmic function. This again has the effect that we're ramping up
the number of accepted loose references before a repack rather fast, but
then put an effective limit on how many refs are accepted in the
repository. The scaling factor of this may need to be adjusted, but
should be a good first approximation of what we need.

Note that we do not take into account how many loose refs there are to
estimate the number of total references: the more refs we have there the
more we want to repack them, especially so if we have no or only a very
small packed-refs file. It thus shouldn't be required to also take those
into accounts.

Changelog: changed