repository: Avoid spawning Git command if there is no worktree to prune
When calling the `Cleanup()` RPC we clean up worktrees that are either
disconnected (the worktree doesn't exist anymore) or stale (the worktree
hasn't been touched for 6 hours). In case of stale worktrees we simply
spawn `git worktree prune` to let Git handle everything for us.

Spawning Git commands is expensive when servers are under high load, so
we should try hard to not do so in case they're not needed. And we can
indeed avoid spawning git-worktree(1) in almost all cases considering
that most repositories are never going to have worktrees in the first
place: we can simply check whether Git's "worktrees" directory contains
any directory, and only if it does we'll spawn the Git command.

Refactor the code to use this heuristic. This is in preparation for
starting to clean up worktrees in OptimizeRepository to avoid needless
busywork.

Changelog: performance