repository: Make maintenance functions more testable
We're about to extend `OptimizeRepository()` to handle all of the
repository's maintenance jobs. It's thus about to become more complex,
which makes it hard to test whether it works alright. This is further
complicated by the fact that most of the maintenance functions we have
right now are receiver functions, which makes them harder to test on
their own, as well.

Refactor the maintenance functions to not be receivers anymore. Most of
them really only required access to a Git command factory to be able to
execute Git commands, and this is trivially to fix by just using ourj
`localrepo.Repo` abstraction instead.