repository: Make repacking of repositories self-contained
While we do have a `repack()` function which handles repacking of
repositories, this function is not self-contained: callers need to pass
additional Git options to request a full repack. This makes it hard to
reuse the logic in other locations where we potentially want to do a
full repakc, as well.

Refactor the code to move all repacking logic into `repack()` itself.
Callers can decide whether they want to do a full or partial repack by
setting the respective parameters in a new config structure.