localrepo: Introduce new functions to get repository paths
The `storage.Locator` interface hosts some of the functions which allow
us to locate specific items in a Git repository. Ultimately, this logic
is Git-specific and shouldn't be hosted in the locator at all.

Prepare for moving it out of the locator interface by providing those
functions via the localrepo package, as well.