storage: Remove Git-specific knowledge from the locator interface
The `storage.Locator` interface provides functions to locate specific
Git data structures in a repository. This knowledge is specific to Git
though and doesn't have anything to do with what the locator should do,
which is to provide information about Gitaly-specific locations.

Move this Git-specific knowledge into the localrepo package and remove
those functions from the interface.