Disable metadata creation hack in object pool service tests
This commit amends the object pool service's tests to not rely on Praefect's
metadata force creation hack. The hack creates metadata for repositories
when Praefect first sees a request for the repository. This has potential
to hide problems in tests and doesn't match the production scenario. To
prepare for removing the hack entirely, this commit refactors the tests
to create the test repositories through the API rather than placing them
directly in the storage.

This commit exposes a bug in Praefect where it does not create metadata
records for the pool repositories as CreateObjectPool is not tagged
handled like the other creation RPCs are. The tests affected by the
bug are skipped.