Handle CreateObjectPool as a repository creation RPC
CreateObjectPool creates the object pool but Praefect does not handle
it as a repository creating RPC. This causes subsequent operation on
the pool repository to fail as Praefect doesn't create the relevant
metadata records. This commit marks the RPC as a repository creating
RPC.

With the metadata correctly created, the object pool service's test
now work with Praefect again without the metadata creation hack. Object
pools still face additional problems on Gitaly Cluster. There's no
special handling from them, so it can be that the pool repository's
primary differs from the main repository's. This may cause problems
as the pool operations are not correctly replicated to the secondaries.
In combination with variable replication factor, it can be that the
pool repository ends up on a different node entirely than the main
repository.

TestStreamDirector had its RPC changed as it's supposed to test a
normal mutator and CreateObjectPool is now handled differently from
most mutators.

Changelog: fixed