repository: Repack after potentially cleaning up corrupted files
In OptimizeRepository we first try to repack the repository if needed
and then perform housekeeping tasks. This order doesn't make a lot of
sense though: housekeeping will also prune any known-corrupt refs which
would potentially break git-repack(1), so if we ran housekeeping first
then chances are higher for the repack to succeed.

Switch the order of both actions to improve our resilience to repository
corruption.