repository: Let OptimizeRepository repack repos without an RPC call
When OptimizeRepository decides to repack the repository it's doing an
RPC call to RepackFull. This is rather pointless though in that it
creates a dependency on the repository service, which makes it harder
than necessary to move the logic into a RPC-independent package at a
later point. Furthermore, it gives us less flexibility in how we repack
the repository if we decide to change the inner workings of
OptimizeRepository.

Remove this dependency by calling `repack()` directly instead of doing
an RPC call.