repository: Factor out code to determine whether repos need repacking
We're about to extend the heuristics used to decide whether a repository
needs repacking or not, which is hard to do right now because heuristics
and repacking are closely coupled right now. Furthermore, the current
layout doesn't easily allow to test the heuristics we have in place.

Make this easier to handle by extracting a function which determines
whether a repository needs repacking or not. This function also returns
a repacking configuration such that it is possible to either perform a
full or incremental repack depending on the repository's on-disk state.

Add tests to verify the function behaves as expected.