repository: Use heuristic for full repacks in OptimizeRepository
Right now we're only repacking repositories in OptimizeRepository when
they're missing auxiliary caches like bitmaps or commit-graphs with
bloom filters. While this has been helpful in rolling out the generation
of these caches, ultimately we're advancing to a state where every repo
should have both of them. As a result, calling OptimizeRepsiotory is
becoming more and more of a no-op over time. What we want is for this
RPC to not only repack repositories if they're lacking crucial caches,
but also when we can determine that they're in an inefficient state.

Introduce a new heuristic which determines whether a repository needs a
full repack. A full repack will collapse all the packfiles we have into
a single packfile, which allows Git to look up objects without having to
search through all packfiles. Furthermore, it allows us to deltify more
objects because deltas can only be created for objects which reside in
the same packfile.

Doing a full repack of the repository scales with the number of objects
which exist in the repository and with their total size: the bigger the
repository is, the longer it takes. As a consequence, we need to become
more careful with repacking the repository as it grows given that we'd
otherwise be repacking it all the time.

To address this we use a heuristic which estimates the repository's size
based on the biggest packfile that exists in the repository. While this
is not an accurate reflection of repository size, it is fast to compute.
From here on we use a logarithmic function to grow the boundary of how
many packfiles need to exist before we do a full repack. This has the
effect that we're initially scaling up the number of allowed packfiles
quite fast, but decellerate when we approach repositories where the
packfiles range in the gigabytes.

This heuristic is imperfect by necessity, but it should give us a first
good iteration of this. Note that this is only a stop-gap solution
anyway: eventually, we want to migrate away from the full/incremental
repack split and always use multi-pack indices and geometric repacking,
which solves this problem of huge repositories in a much saner way.

Changelog: changed