repository: Log what optimizations where applied to a repo
We're about to extend OptimizeRepository's functionality to not only
cover packfiles, but also to repack references. Because we have no
logging in place right now and because we'll use heuristics to determine
whether we need to repack or not it's hard to know what exactly has been
optimized.

Introduce logging into OptimizeRepository that gives some insight into
what we've been doing.