migrate: Add -verbose flag to sql-migrate
Currently the only way to determine the statement being executed by a
migration is to check the Gitaly source. Administrators and Support may
not have the Gitaly repo on-hand, making it slow and inconvenient to
check what a problem migration is doing.

Let's give admins an easy way to review the changes being performed by a
migration by adding a `-verbose` flag to the `praefect sql-migrate`
subcommand that prints the migration statement.

Example output:

```
=  2022-02-01T16:00:56 20200109161404_hello_world: migrating
    [CREATE TABLE hello_world (id integer)]
== 2022-02-01T16:00:56 20200109161404_hello_world: applied (9.525348ms)
```

Changelog: added