Send `git fetch` details in FetchRemote response
When issuing a FetchRemote RPC during a project import, the RPC fails
if tags need to be updated unless the force option is specified. It
would be helpful to the caller to be able to know which tags actually
needed to be updated. This commit adds a new protobuf option,
`send_fetch_status`, that will send back a structured response to the
caller if requested.

Relates to https://gitlab.com/gitlab-org/gitaly/-/issues/4033

Changelog: added