Remove metadata creation hack from Praefect
Praefect contains a testing hack that created metadata for repositories
when Praefect first received a request for them. This allowed Praefect
to be used in Gitaly tests that placed repositories directly in the
storage rather than creating them through the API. Placing the
repositories directly in the storage didn't allow for Praefect to
create the required metadata entries on repository creation which in turn
caused failures in the tests as Praefect would return a not found. All of
the said tests have been changed to create the repositories through the
API, so the hack is no longer needed. This commit removes the hack from
Praefect and the associated configuration options.