Propagate NotFound error returned by GetRepoPath
We are currently wrapping some of the errors returned by GetRepoPath,
which makes them lose the gRPC NotFound code.

This leads to a gRPC error code of Unknown, which counts towards our
error SLO, paging the on-call.

This patch removes the wrapping, ensuring the error code is propagated
to the client, and accounted for appropriately in logs and metrics.

Changelog: fixed