gitaly-git2go: Send all errors over gob
Previously stderr was used to send plaintext errors back to the caller.
Since we are looking to use stderr for logging, we need to stop sending
errors this way. Here we use the universal gob response error field
instead.

Client side must continue reading errors from both gob and stderr for a
release so that we can safely upgrade.