Move pool path validation logic to housekeeping package
Object pool paths are validated to match a regex of the pool paths
Rails is generating. As Praefect will soon be generating the disk
paths, the path of the pools will no longer match what Rails is
generating with the @pools/xx/xx/xxxx format. Pool paths are also
verified in the housekeeping package to avoid pruning objectp pools.
This commit moves the pool path checking logic to the houskeeping
package so it can be shared between the objectpool and housekeeping
packages. It's not possible to locate the shared functionality in
git nor the objectpool packages due to cyclic dependencies between
them. With the check in a common place, a follow up commit can later
add validation regexp for Praefect generated pool paths.